+++ client/src/modules/Registry/components/Disciplines/Disciplines.test.tsx
@@ -2,7 +2,7 @@ import { render, screen } from '@testing-library/react';
 import { Form } from 'antd';
 import { DisciplineDto } from 'api';
 import { LABELS } from 'modules/Registry/constants';
-import { Disciplines } from './index';
+import { Disciplines } from './Disciplines';
 
 const mockDisciplines = [
   {
