+++ client/src/modules/Registry/components/Footer/Footer.tsx
