+++ client/src/modules/Registry/components/Form/DoneSection/DoneSection.tsx
