+++ client/src/modules/Registry/components/PersonalInfo/PersonalInfo.test.tsx
@@ -1,7 +1,7 @@
 import { cleanup, fireEvent, render, screen } from '@testing-library/react';
 import { Form } from 'antd';
 import { ERROR_MESSAGES, LABELS, PLACEHOLDERS } from 'modules/Registry/constants';
-import { PersonalInfo } from './index';
+import { PersonalInfo } from './PersonalInfo';
 
 const mockValues = {
   firstName: 'John',
