+++ client/src/modules/Registry/components/index.tsx
@@ -1,3 +1,10 @@
-export * from './Form/DoneSection';
-export * from './Form/GeneralSection';
-export * from './Form/MentorshipSection';
+export * from './Form/DoneSection/DoneSection';
+export * from './Form/GeneralSection/GeneralSection';
+export * from './Form/MentorshipSection/MentorshipSection';
+export * from './PersonalInfo/PersonalInfo';
+export * from './AdditionalInfo/AdditionalInfo';
+export * from './ContactInfo/ContactInfo';
+export * from './Disciplines/Disciplines';
+export * from './Footer/Footer';
+export * from './FormCard/FormCard';
+export * from './Preferences/Preferences';
