+++ client/src/modules/Registry/pages/Mentor/index.tsx
@@ -4,8 +4,7 @@ import { Location } from 'common/models';
 import { RegistrationPageLayout } from 'components/RegistartionPageLayout';
 import { Session } from 'components/withSession';
 import css from 'styled-jsx/css';
-import { MentorshipSection, DoneSection, GeneralSection } from 'modules/Registry/components';
-import { Footer } from 'modules/Registry/components/Footer';
+import { MentorshipSection, DoneSection, GeneralSection, Footer } from 'modules/Registry/components';
 import { useCallback, useEffect, useState } from 'react';
 import { useAsync, useUpdate } from 'react-use';
 import { CdnService } from 'services/cdn';
